requirements.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
scipy>=1.10.0

# Embeddings and transformers
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# Dimensionality reduction and visualization
umap-learn>=0.5.0
plotly>=5.14.0

# Graph analysis and community detection
networkx>=3.0
python-igraph>=0.10.0
leidenalg>=0.9.0
pyvis>=0.3.0

# Keyword extraction
keybert>=0.8.0

# API communication (for Ollama)
requests>=2.28.0

